I've seen some things
and so I'm making another pass.

Modified the goals and added some clarifications to what they meant. Took
another pass as the type system, I'm trying to see where to land on this.
Debating on going full Hindley-Milner or some weird mutant... :shrug:

I also did some spot corrections to the example codes as well as tryed to make
things consistent across pages

Also I'm ditching the blocks and trying a whitespace sensitive approach so
there's some of that here in this commit as well